mini project - iNaturalist
This purpose of this project was to integrate the iNaturalist API into the FC website. Visitors to forests would upload pictures of the wildlife they took for others to see. A pin would be registered for each upload. Users would be able to click on the pins to see what was uploaded, what time and where.

The images show a before and after on both desktop and mobile devices
the twig template was taken from iNaturalist and edited. I coded up the scss and js files to respond and animate the map information once a user interacts with the pins