jQuery examples
i developed the radio button and checkbox functionality using jQuery. This is in the dev and staging environment ready for feedback before going live.

VI-bar-desktop-current shows the current VI functionality on the site right now. The other VI-bar files shows what will be released in the next few weeks. i was tasked with coming up with a way to make the Visitior Information bar more prominent. jQuery was used for the accessibility and scss was used to better display the VI area and show prominence